Check for existing file

+++ secdownload.py
@@ -76,7 +76,10 @@ with canvas_session() as s:
                     continue
                 if numatt != 1:
                     print(f'{stu["name"]}: {numatt} attachments, ?')
-                dst = f'{codename(stu)}_{sid}_' + sub['attachments'][-1]['filename']
+                dst = Path(f'{codename(stu)}_{sid}_' + sub['attachments'][-1]['filename'])
+                if dst.exists():
+                    print(dst, 'found')
+                    continue
                 url = sub['attachments'][-1]['url']
                 print('Fetching', dst, '... ', end='', flush=True)
                 with s.get(url, stream=True) as fileget, open(dst, 'wb') as out:
